Print receipt all format completed

+++ app/src/main/java/com/lkintechnology/mBilling/activities/company/navigations/TransactionPdfActivity.java
@@ -9,12 +9,13 @@ import android.graphics.drawable.ColorDrawable;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;
-
 import android.os.StrictMode;
 import android.os.Handler;
 import android.print.PdfPrint;
 import android.print.PrintAttributes;
 import android.print.PrintDocumentAdapter;
+import android.support.design.widget.CoordinatorLayout;
+import android.support.design.widget.Snackbar;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -28,10 +29,27 @@ import android.view.View;
 import android.webkit.WebView;
 import android.widget.TextView;
 import android.widget.Toast;
-
-
 import com.lkintechnology.mBilling.R;
+import com.lkintechnology.mBilling.activities.app.ConnectivityReceiver;
+import com.lkintechnology.mBilling.activities.company.CompanyListActivity;
+import com.lkintechnology.mBilling.entities.AppUser;
+import com.lkintechnology.mBilling.networks.ApiCallsService;
+import com.lkintechnology.mBilling.networks.api_response.purchase_return.GetPurchaseReturnVoucherDetails;
+import com.lkintechnology.mBilling.networks.api_response.purchase_return.PurchaseReturnVoucherDetailsData;
+import com.lkintechnology.mBilling.networks.api_response.purchasevoucher.GetPurchaseVoucherDetails;
+import com.lkintechnology.mBilling.networks.api_response.purchasevoucher.PurchaseVoucherDetailsData;
+import com.lkintechnology.mBilling.networks.api_response.sale_return.GetSaleReturnVoucherDetails;
+import com.lkintechnology.mBilling.networks.api_response.sale_return.SaleReturnVoucherDetailsData;
+import com.lkintechnology.mBilling.networks.api_response.salevoucher.GetSaleVoucherDetails;
+import com.lkintechnology.mBilling.networks.api_response.salevoucher.SaleVoucherDetailsData;
+import com.lkintechnology.mBilling.utils.BluPrinterHelper;
+import com.lkintechnology.mBilling.utils.Cv;
+import com.lkintechnology.mBilling.utils.Helpers;
+import com.lkintechnology.mBilling.utils.LocalRepositories;
 import com.lkintechnology.mBilling.utils.TypefaceCache;
+
+import org.greenrobot.eventbus.EventBus;
+import org.greenrobot.eventbus.Subscribe;
 
 import java.io.File;
 import java.util.List;
@@ -40,15 +58,24 @@ import butterknife.Bind;
 import butterknife.ButterKnife;
 
 public class TransactionPdfActivity extends AppCompatActivity {
-
+    @Bind(R.id.coordinatorLayout)
+    CoordinatorLayout coordinatorLayout;
     @Bind(R.id.webView)
     WebView mPdf_webview;
+    Snackbar snackbar;
+    ProgressDialog mProgressDialog;
+    public SaleVoucherDetailsData dataForPrinterSale;
+    public SaleReturnVoucherDetailsData dataForPrinterSaleReturn;
+    public PurchaseVoucherDetailsData dataForPrinterPurchase;
+    public PurchaseReturnVoucherDetailsData dataForPrinterPurchaseReturn;
 
     private static final String TAG = "PdfCreatorActivity";
     private File pdfFile;
     final private int REQUEST_CODE_ASK_PERMISSIONS = 111;
     String htmlString;
     ProgressDialog progressDialog;
+    AppUser appUser;
+    String type;
 
     @TargetApi(Build.VERSION_CODES.CUPCAKE)
     @Override
@@ -59,6 +86,7 @@ public class TransactionPdfActivity extends AppCompatActivity {
         progressDialog = new ProgressDialog(TransactionPdfActivity.this);
         progressDialog.setMessage("Please wait...");
         progressDialog.show();
+        appUser = LocalRepositories.getAppUser(this);
         new Handler().postDelayed(new Runnable() {
             @Override
             public void run() {
@@ -68,6 +96,12 @@ public class TransactionPdfActivity extends AppCompatActivity {
                 initActionbar();
                 Intent intent = getIntent();
                 String company_report = intent.getStringExtra("company_report");
+                if (intent.getStringExtra("type")!=null){
+                    type = intent.getStringExtra("type");
+                }else {
+                    type = "else";
+                }
+
                 htmlString = company_report;
                 Spanned htmlAsSpanned = Html.fromHtml(company_report);
                 htmlString = htmlAsSpanned.toString();
@@ -76,9 +110,22 @@ public class TransactionPdfActivity extends AppCompatActivity {
             }
         },3*1000);
 
+        EventBus.getDefault().register(this);
+
 
     }
 
+    @Override
+    protected void onStop() {
+        EventBus.getDefault().unregister(this);
+        super.onStop();
+    }
+
+    @Override
+    protected void onPause() {
+        EventBus.getDefault().unregister(this);
+        super.onPause();
+    }
 
     private void initActionbar() {
         ActionBar actionBar = getSupportActionBar();
@@ -159,15 +206,50 @@ public class TransactionPdfActivity extends AppCompatActivity {
     public boolean onOptionsItemSelected(MenuItem item) {
         int id = item.getItemId();
         if (id == R.id.icon_id) {
-
-            File path = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM + "/m_Billing_PDF/");
-            if (path.isDirectory()) {
-                String[] children = path.list();
-                for (int i = 0; i < children.length; i++) {
-                    new File(path, children[i]).delete();
+            if (CompanyListActivity.boolForInvoiceFormat){
+                Boolean isConnected = ConnectivityReceiver.isConnected();
+                if (isConnected) {
+                    mProgressDialog = new ProgressDialog(this);
+                    mProgressDialog.setMessage("Info...");
+                    mProgressDialog.setIndeterminate(false);
+                    mProgressDialog.setCancelable(true);
+                    mProgressDialog.show();
+                    if (type.equals("sale_voucher")){
+                        ApiCallsService.action(getApplicationContext(), Cv.ACTION_GET_SALE_VOUCHER_DETAILS);
+                    }else if (type.equals("sale_return_voucher")){
+                        ApiCallsService.action(getApplicationContext(), Cv.ACTION_GET_SALE_RETURN_VOUCHER_DETAILS);
+                    }else if (type.equals("purchase_voucher")){
+                        ApiCallsService.action(getApplicationContext(), Cv.ACTION_GET_PURCHASE_VOUCHER_DETAILS);
+                    }else if (type.equals("purchase_return_voucher")){
+                        ApiCallsService.action(getApplicationContext(), Cv.ACTION_GET_PURCHASE_RETURN_VOUCHER_DETAILS);
+                    }
+                    System.out.println("");
+                } else {
+                    snackbar = Snackbar
+                            .make(coordinatorLayout, "No internet connection!", Snackbar.LENGTH_LONG)
+                            .setAction("RETRY", new View.OnClickListener() {
+                                @Override
+                                public void onClick(View view) {
+                                    Boolean isConnected = ConnectivityReceiver.isConnected();
+                                    if (isConnected) {
+                                        snackbar.dismiss();
+                                    }
+                                }
+                            });
+                    snackbar.show();
                 }
+
+            }else {
+                File path = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DCIM + "/m_Billing_PDF/");
+                if (path.isDirectory()) {
+                    String[] children = path.list();
+                    for (int i = 0; i < children.length; i++) {
+                        new File(path, children[i]).delete();
+                    }
+                }
+                createWebPrintJob(mPdf_webview);
             }
-            createWebPrintJob(mPdf_webview);
+
         } else {
             finish();
         }
@@ -178,6 +260,58 @@ public class TransactionPdfActivity extends AppCompatActivity {
     public void onBackPressed() {
         finish();
         super.onBackPressed();
+    }
+
+    @Subscribe
+    public void getSaleVoucherDetails(GetSaleVoucherDetails response) {
+        mProgressDialog.dismiss();
+        if (response.getStatus() == 200) {
+
+            dataForPrinterSale = new SaleVoucherDetailsData();
+            dataForPrinterSale = response.getSale_voucher().getData();
+            BluPrinterHelper.saleVoucherReceipt(this,dataForPrinterSale);
+        } else {
+            Helpers.dialogMessage(TransactionPdfActivity.this, response.getMessage());
+        }
+    }
+
+    @Subscribe
+    public void getSaleReturnVoucherDetails(GetSaleReturnVoucherDetails response) {
+        mProgressDialog.dismiss();
+        if (response.getStatus() == 200) {
+
+            dataForPrinterSaleReturn = new SaleReturnVoucherDetailsData();
+            dataForPrinterSaleReturn = response.getSale_return_voucher().getData();
+            BluPrinterHelper.saleReturnVoucherReceipt(this,dataForPrinterSaleReturn);
+        } else {
+            Helpers.dialogMessage(TransactionPdfActivity.this, response.getMessage());
+        }
+    }
+
+    @Subscribe
+    public void getPurchaseVoucherDetails(GetPurchaseVoucherDetails response) {
+        mProgressDialog.dismiss();
+        if (response.getStatus() == 200) {
+
+            dataForPrinterPurchase = new PurchaseVoucherDetailsData();
+            dataForPrinterPurchase = response.getPurchase_voucher().getData();
+            BluPrinterHelper.purchaseVoucherReceipt(this,dataForPrinterPurchase);
+        } else {
+            Helpers.dialogMessage(TransactionPdfActivity.this, response.getMessage());
+        }
+    }
+
+    @Subscribe
+    public void getPurchaseReturnVoucherDetails(GetPurchaseReturnVoucherDetails response) {
+        mProgressDialog.dismiss();
+        if (response.getStatus() == 200) {
+
+            dataForPrinterPurchaseReturn = new PurchaseReturnVoucherDetailsData();
+            dataForPrinterPurchaseReturn = response.getPurchase_return_voucher().getData();
+            BluPrinterHelper.purchaseReturnVoucherReceipt(this,dataForPrinterPurchaseReturn);
+        } else {
+            Helpers.dialogMessage(TransactionPdfActivity.this, response.getMessage());
+        }
     }
 }
